Welcome survey: store which group a user is in
After an account is created, figure out which group a user
is in and store it immediately so we have this information
available if the user never clicks the save of skip
buttons of the survey.

Refactor all the code related to figuring out the experiment,
the questions, and handling the responses out of the special
page into a 'WelcomeSurvey' class so it can be re-used from
hooks.

Use a random number to decide which group a user should
be in instead of the last digit of the user id. See
discussion in task for rationale.

Generally rename 'experiment' to 'group', it's more
accurate.

Bug: T206371
Change-Id: I3fcae6d62b0fd2fc9cdbf918748412d5cc5437cc