Don't assume a POSTed survey shows the confirmation page
It is rare but possible to submit the survey form and get the survey
back (with onSubmit() never called). The only case I know is when the
form stays open for too long and the token needs to be reviewed.

This change ensures that the page title is set to the confirmation
message only when we are sure to show the confirmation page.

Bug: T206371
Change-Id: Ic46a826f2cec14a713c586dc75e11926f9eb4665