Help panel: Move config vars into a data module
If the decision to load the help panel module is made dynamically on
the client side, rather than on the server side, we need to be able to
pull in the config vars along with the module.

The config var with the user's email can't be loaded this way, because
non-embedded RL modules don't have access to the User object. So
instead, export the email config var on any page view that could
conceivably display a help panel (even if it's not action=edit).

Bug: T211059
Depends-On: I38ece7320adab64113c5d2a91056d5522f369367
Change-Id: I91650d04bdb89806b3228217e79eeceaff3b48fd