Help Panel: Use an init module to conditionally add a VE plugin module
We were previously unconditionally loading the HelpPanel module as a VE
plugin module, and the help panel appeared on every VE load even if it
was disabled in config. We could fix that by checking the config at the
start of the HelpPanel module, but then we've already wastefully loaded
the entire module.

We could also make the VE plugin module an init module that checks the
config, then loads the rest of the code if the help panel should be
shown, but that requires a second request.

Instead, add an init module on page views where we think we may show a
help panel later. This module then registers the HelpPanel module as a
VE plugin, so it's loaded alongside VE in the same request.

(Also add a config check at the top of the main HelpPanel module for
paranoia.)

Change-Id: I21c39c134010ee0a64b302b8f89d273e8f23e86b