Help Panel: Don't autofocus on open
Without this patch, if localStorage is empty and you open the help panel, your
eyes are drawn to a red outline around the text input. This implies an error
which is confusing to the end user, since the user is welcome to interact with
the help panel without posting a question.

With the patch, no red outline is drawn on panel open. If the user clicks in
the text area then out of it again, the red outline will be drawn.

Change-Id: Ic8de01897a003b1d89a3bca1e13aa233a9ff1add