Help panel: Animate the CTA button onto the screen
Most of this change involves wrapping the CTA button in a <div>, and
moving all the positioning CSS (as well as the animation) onto that
<div>. That way we can hide and show the button separately from its
wrapper. Without this, opening and closing the popup would cause the
button to animate in again (because the animation happens again when the
button is un-hidden).

Bug: T212887
Change-Id: I5054c68a823bbf44298734456b15c068e0ea5ef4