Help panel: Prevent iOS from zooming in on text input
iOS will zoom when focusing text inputs, but only if their font size is
less than 16px. Work around this by making the font size in the help
panel be 16px on mobile (while keeping it at 14px on desktop).

Also try to make the behavior of the text input in the footer a bit less
crazy by undoing the unselectability rules on the footer.

Bug: T212505
Change-Id: Idb861ba7a08cc6a2f409f373161620c85a872272