Support templates and magic words in wgGEHelpPanelHelpDeskTitle
kowiki's help desk is organized as a set of subpages of
the main help desk page, one for each month. The exact name
of the subpage is generated by a template ({{Y-M}})

This change allows configuring wgGEHelpPanelHelpDeskTitle
with the same template so the help panel posts the
questions to the correct page.

Bug: T213186
Change-Id: I0e9b50fb08d596d70e19f0db7e19203681a8982a