Help panel: ResourceLoaderHelpPanelModule handle help panel disabled
Because ResourceLoaderHelpPanelModule has versioning based on content,
it's getScript() method gets called (to build a version based on the
content's hash) even when the help panel is disabled and not configured.
In this case the getScript() method errors out and prevent other
features (Welcome survey) in this extension from working.

This patch proposes to keep versioning based on content but
return empty string when the help panel is disable.

Also remove default values that don't make sense for
wgGEHelpPanelHelpDeskTitle and wgGEHelpPanelViewMoreTitle.
These config values were hidding the issue locally.

Bug: T213356
Change-Id: I30b3da4553178cfb658cafddb43658f01c0a166c