Help panel search: hide help content when searching
It used to be that the help panel would replace the help
content as soon as the user focuses the search input.

With this patch, nothing happens on focus but as soon as
the user starts typing, the help links are hidden
and search results (or no-results message) are shown.

Bug: T209301
Change-Id: I842508111cea7a79e29575f5da76baec8e64ca83