Help Panel: Fix iOS scroll bug
Adjust detach/append hack to apply to any MobileFrontend overlay, not just
editor-overlay. This includes the loading overlay which at the moment is not
always hidden, and interferes with scroll.

In addition, detach/append the VE UI overlay as this also interferes with
scrolling the search results.

Finally, remove the 'scrollable: true' attempt at fixing this earlier which
didn't have a discernible impact.

Bug: T215430
Change-Id: I98fb132a0fc25bb4d3418a00479274ce9327ad46