Help Panel: Fix IME broken in help panel search
As noted in the bug report in T216131, our use of two input fields and calling
setValue as we change from the first to the second breaks IME. Specifically, the
first character entered in the home panel search input field is ignored by IME
in the second input field, although the user cannot visibly see this. As a
result, word composition is broken in languages like Korean.

To fix this we are rolling back to the UX we had in place prior to
I842508111cea7a79e29575f5da76baec8e64ca83, where focusing on the search field
swaps to the search panel. With this UX there is no need to call setValue since
the user begins typing in the search panel input field.

Not all code has been reverted (e.g. code related to the clearsearch action) but
that can happen in a later point in time, depending on what our longer term
fix for this bug is.

Bug: T216131
Change-Id: I14bb08f6bea47dd9e7163a31a776b27117beb5dd