Help panel: Conditionally depend on mediawiki.ForeignApi when needed
On desktop, mw.ForeignApi is loaded by Echo, so using it in
HelpPanelSearchWidget works by accident. But on mobile, it's not loaded,
and trying to use it causes a JS error.

I didn't want to add mw.ForeignApi as a dependency, because its use is
disabled in production. Instead, use mw.loader.using() to load it when
we need it, and put some ugliness in place to work with the fact that
this.api is created asynchronously.

Bug: T217218
Change-Id: I2544cbef6ca212b7fffe7fbb47a3ea5717f9f396