Fix reading mode detection
wgAction can't be trusted because in some cases it is
possible to toggle between reading and edit with VE
without a page reload. Therefore, wgAction is often
not up to date with the current mode.

This patch uses uri.fragment to detect reading mode
on mobile and falls back to 'reading' on desktop
when no known editor is detected.

It also fixes a bug where the help panel would stay open
if it is open when we go back to reading mode from VE
without a page reload.

Bug: T217802
Change-Id: I0c86af7f2306d5ec1fb16e01fc676436aa96d35b