Refactor helppanel code
The basic idea is to separate generic "ask a question" feature
from its use in specific contexts (help panel, help module).

ext.growthExperiments.Help: this module contains
reusable components that can be used to ask a
question (actually post a new section) on a page.

ext.growthExperiments.HelpPanel: this module contains
the CTA and uses the module above to create the
help panel.

Change-Id: I22aeae8121ff9a0972f96d41115a53655df34978