Fix desktop reading mode detection
Desktop reading mode was assumed if no known editor
was found. It turns out not to work when loading
a page with ?action=edit and the default editor
is VE. In this case, loading the editor takes a bit
of time so when 'HelpPanel.cta.js' file is executed,
it fails to find the editor.

This patch checks the 'action' and 'veaction' query
parameters to know if an editor might be loading.

Bug: T217802
Change-Id: Ib1b059124fb6d3a8a9285e54199a6ffd75f219b9