Log 'close' only on user-initiated close
The 'close' event was logged every time the help
panel was closed, even when the panel is closed
by the software during a change of context.

With this patch, the 'close' event is only
logged when the user clicks on one of the
close buttons.

Change-Id: Ib13cf88d182881e652fad168dc7da6921363e6ba
(cherry-picked from commit 7b4eefb977730b08f4f337602c5d491e8cc94447)