Provide configurable header and subheader tags
By default, we can render the header and subheader as h2 and h3, but for the
start modules it would make sense for the header to be an h3. AFAICT there is no
subheader for the start modules, but adding it as a possible override alongside
the header tag makes sense as we may use it sometime in the future.

Change-Id: I93d2eff106402c7c6d75e6d2a46f8b76308d70c0