Homepage layout and styling
Use a flexbox-based column layout for the page: a 75%/25% split between
the main column and the sidebar column, with the sidebar containing the
help module and the main column containing everything else. The entire
first line of the main column is taken up by the start module, and the
second line is split 50%/50% between the impact and mentorship modules.

Also:
* Styling for the help and mentorship modules
* Add box shadows
* Bring back bottom borders for headings and space them out properly
* Combine edit count and last active in one line in mentorship module

Bug: T218966
Change-Id: I8f9a98f96d2d45e9cc40e07f032ab3b2dbcb8609