Homepage: Fix escaping of things passed to buildSection()
BaseModule::buildSection() documents its $content parameter as "text or
HTML", and the return values of getHeader(), getBody() etc are also
documented as "text or HTML". But there is no such thing, because text
must be escaped. Passing plain text for a parameter where HTML is
expected is dangerous.

Explicitly document these parameters and return values as HTML not text,
and escape everything (mostly messages) that ends up there.

Change-Id: I944d3f60c1903e238742dcfdd5b05abbbfb7c196