Remove tutorial visit job
In retrospect, we can log the tutorial visits through a simpler mechanism than
the job queue while also avoiding transaction profiler expectation violations.
This patch reworks the tutorial button to submit a POST request to a subpage
of Special:Homepage, save the user pref, then redirect onwards to the tutorial
title.

The business logic in this commit changes slightly as random visits to the
configuration-defined Tutorial title won't result in the module showing as
completed, however the behavior in this patch still aligns with the overall
intent of this feature, which is to document that a user clicked the tutorial
button and visited the tutorial page.

Change-Id: I1275296586762e8150549341c43c10040f04e6ea