Remove mentor save job
Assign a mentor in the LocalUserCreated hook (if the user is in the experiment
group) or from the UserSaveOptions hook (if the user has enabled the homepage in
their preferences). Both contexts are POST and will not trigger transaction
profiler warnings.

This is a slight change in business logic since visiting the Special:Homepage
without being part of the experiment group or enabling the homepage pref will no
longer assign a mentor.

Change-Id: Ieea97814ea65a65ba89b95718597f40657f19eec