Add user email vars for Help and Mentorship modules
The Help and Mentorship modules reuse part of the Help Panel
to post questions to the help desk or a mentor's user talk page,
respectively. The "question poster" feature they use relies
on 2 config variables to correctly handle the user's email
state (none, unconfirmed, confirmed) but those variables
were only added to the page by the HelpPanelHooks when the
help panel is likely to be shown.

This patch moves the definition of those variables to HelpPanel
and adds them to the page from the three places that might need
them.

While they may be added to the output up to 3 times per request,
the OutputPage class keeps a single list of config variables
keyed by config name so they won't be duplicated.

Bug: T221810
Change-Id: I12510b81dfbe23091de6c2c20087537a14c236c5