Homepage: Remove footers for start module submodules
Move the "guidelines" link in the userpage submodule from the footer
to the subheader, and delete the "change email" link in the footer
of the email submodule altogether.

This means none of the start module submodules have footers anymore, and
the subheaders have gotten taller, so adjust the min-height percentage.

Bug: T217105
Change-Id: I170271a3be2624e947ca8dc1d0cc5495d64a9971