Drop RENDER_NOW flag, use thumbnail script for impact images
In profiling a Special:Homepage with an impact module that has three images
displayed, it appears that File::transform takes several hundred
milliseconds. This is because we pass the RENDER_NOW flag to the transform
method on the image file.

By dropping RENDER_NOW, the background image URL will utilize thumb.php,
postponing image loading (and file transformation if necessary) until after the
page has been delivered to the client.

Bug: T222223
Change-Id: I5d0c8d35caa8c133ef4730c082a39c5a4950b01b