Homepage: Use flexbox to align buttons in start module
Make the "subheader" (which should really be called the body) take up
all the remaining space, and make the header and the footer take only
the space they need. This aligns the tops of the headers, and the
bottoms of the buttons. If one of the buttons line wraps, this looks a
little odd, but it's a much better solution for aligning them than the
min-height stuff we were using before. This also means that line
wrapping buttons can't spill out of the container anymore.

Also fix the header c, by using a div instead of a span
for the header text and setting overflow: visible;. After doing both of
these things, we don't need min-height anymore.

Change-Id: I16e6a27d0530954a0a8784ecf0b61a1916ba6372