Homepage: Move start module subheaders to bodies, bodies to footers
Start module submodules were using the subheader for the bulk of their
content, the body for the buttons at the bottom, and weren't using the
footer at all. Instead, put the bulk of the content in the body, put the
buttons in the footer, and don't use the subheader. This makes the
naming more sensible, and also makes the other submodules consistent
with the account submodule, which doesn't have a button and was already
using the body instead of the subheader.

Also use <h3>s instead of <h2>s for start submodule headers, and a <div>
instead of an <h3> for the impact module subheader.

Change-Id: I9851b0f42c5d341e7e69e82b7117d099e976e5be