Prevent parsing of GEHelpPanelHelpDeskTitle from accessing the session
GEHelpPanelHelpDeskTitle can be configured to use a template. When it is
parsed, it should not try to access the session since it is used
from a RL data module.

I would love to find a way to test that the session is NOT accessed.
Suggestions are welcome. Could be in a follow-up.

Bug: T224224
Change-Id: I291831adbd0380187de31ee947628ae07139f367