Help module: remove weird left margins from (archive) links
There was a CSS rule in the help module that applied a left margin to
all links inside lists in the help module. This was intended for the
list of help links, but it also ended up applying to the list of recent
questions, and to the nested (archive) links inside those. So we had
links in the middle of a sentence getting a small left margin applied to
them and looking really weird, like "Foo ( archived)".

To avoid this, add a class to the help links list and only apply the CSS
rule to that list.

Change-Id: Iafb1d349bac8079982f6c92874e13574a7771cc2