Homepage: Open ask question dialog in route
Define a route (`#/homepage/{moduleName}/question`) and navigate to it when the
homepage question asking modules (help, mentor) are clicked. This allows back
button behavior to correctly return to the previous location for the MF
overlay. The route is also used on desktop as it simplifies the code, and
does not otherwise impact the user experience.

A similar implementation should be done for the help panel, but this patch
restricts the scope of the fix to the homepage modules for now.

Bug: T226041
Change-Id: I50047fdf15453e638b337a447bad28d2a26cec5c