Show pending icon if no pageviews exist for any edit
In order to know if we should show the clock icon we need to have run the
calculations used to generate the edits table, which isn't cheap. But from
looking at the profiler it's not terribly expensive either. To help compensate
for the additional overhead, this patch also reuses instantiated classes when
outputting data for the mobile overlay details.

Bug: T225560
Change-Id: I76146ceff8bbf09fa253060be9a073f424f6782e