Add link to Special:Homepage on Special:Contributions
Only when the user is viewing their own contributions, and has no
contributions. We can't directly tell whether the contributions list
will show an empty result, so we use editcount===0 as a proxy.

MobileFrontend intercepts Special:Contributions and redirects it to
SpecialMobileContributions, which doesn't have a corresponding hook.
Instead, hook into SpecialPageAfterExecute, add our HTML if the special
page being executed is SpecialMobileContributions, and use CSS to hide
the "The filter provided no history matches" message.

Bug: T225328
Bug: T227575
Change-Id: I263765f3090e80511e427089d89e0bcfca6331e9