Mentor: Fix parsing of empty mentor description
If you had a mentors list that looks like:
    * [[User:Alice]]|
    * [[User:Bob]]|Hi, I'm Bob
then the mentor description for Alice should be an empty string, but was
instead the entire like below ("* [[User:Bob]]|Hi I'm Bob").

This was because the regex for parsing mentor descriptions used \s,
which matches the line break immediately after the pipe. To avoid this,
use \h, which only matches horizontal whitespace, and doesn't match
vertical whitespace.

Bug: T227535
Change-Id: I882dd739fed17dcf8028b8dc59cc3a2efa2a5650