Help panel: Fix cog popup alignment
Set the popup alignment to 'backwards', so that it is aligned that way
on purpose rather than by accident. (It was trying to center itself,
then being pushed to the left to stay within the dialog boundaries.)

Update the margin adjustment accordingly, and add a few pixels so that
the border of the cog popup coincides with the border of the dialog.

Bug: T225669
Change-Id: I7d902e1db354f529f037de58c0efc5c1a41581fa