Homepage: Toggle overflow submenu option
This patch makes use of a patch to Minerva (see Depends-On) which results in the
relevant title set in Special:Homepage to be the title used in Minerva's
UserPageHelper class. That tells Minerva that when the user accesses
Special:Homepage, Minerva should render the page as a UserPage. This results in
the tagline getting output.

The hack is that OPTION_OVERFLOW_SUBMENU needs to be set to true when the user
is on Special:Homepage in order to hide the various controls for
editing/watching a page.

Depends-On: I4c1add98167ae908a05b2224adea0c417eb4a290
Bug: T225663
Change-Id: I20cce5bd58cdfbf21c0917905df15ee1f36e68d1