Homepage: Fix missing icon for start module when complete
When all four submodules of the start module are completed, the icon for
the start module disappears, because it gets inverted. Prevent this by
overriding shouldInvertHeaderIcon() to always return false.

This only broke on desktop. On mobile, it wasn't broken because
Start::getMobileSummaryHeader() didn't check shouldInvertHeaderIcon()
and always passed in false. Now that shouldInvertHeaderIcon() is fixed,
the override of getMobileSummaryHeader() is identical to the grandparent
class's version, so call it directly instead.

Bug: T229515
Change-Id: I74d3feb2972a2f8383acda88e0ddd49ea4cd5988