Impact module: Put page view count in its own column in mobile summary
The alignment of the big blue pageviews number was weird in cases where
the "Views for recent edits" text to the left of it wraps (because its
translation is long, or the screen is narrow). One of the reasons this
isn't easy to fix is because the "Views from recent edits" text and the
big number are in a column layout together, but the number is typically
taller than the text. Hacks were added to prevent this from causing
empty space to appear above the text, but those hacks also make it more
difficult to get the alignment to behave the way we want.

This change instead makes the entire module a two-column layout, with
all the text being in the left column and the number in the right
column. This works a lot better, because the number is not as tall as
the 2-3 lines of text combined. It does mean that, if the number is
wide, all the text in the module will wrap, not just the text
immediately beside it.

Bug: T227104
Change-Id: I289cc51fc12181cb8939638cef362f0ea390c4d0