Homepage: Restructure mobile overlay handling
When opening a question dialog, we navigate to #/homepage/help/question
(or #/homepage/mentorship/question), and the factory function for that
route generates an overlay with the module name 'help/question' and no
module data. Miraculously, this doesn't crash, but just puts "undefined"
everywhere, including in the overlay title bar. This "undefined" text is
briefly visible when the dialog opens, and permanently visible if the
screen is large enough that the dialog isn't full-screen.

Instead, cache the overlay objects after constructing them, and return
the same object again when navigating to the question dialog. This means
the overlay doesn't get hidden or switched when the question dialog is
opened, and it continues to be visible underneath the dialog. Caching
overlay objects also prevents them from being wastefully reconstructed
from scratch every time they're opened.

When navigating between #/homepage/help and #/homepage/help/question,
the OverlayManager hides the overlay then immediately shows it again.
This isn't visible to the user, but it does mean we can't listen to the
'hide' event to instrument when the user truly hides the overlay.
Instead, do this from a 'route' event handler that tracks which module
is currently active and logs when that changes. This is a little
complicated, but it's still less complicated than what we were doing
before.

We also have to move the impression logging code: impressions were being
logged when the overlay was built, and that worked because we rebuilt
the overlays each time (and not logging impression events when closing
dialogs worked by accident, due to the stack feature in OverlayManager).
Instead, do impression logging in the 'route' event handler.

Because the 'start' module is weird, we also have to find submodules
in each module and log impressions for them. Impression events for the
'start' module itself are ignored, and the instrumentation expects
impression events for each submodule instead.

Bug: T229271
Depends-On: Idac6553789ff224e3584f730f9b657f295e2052f
Change-Id: If529f60e99f06d4e3db1f53562f0de5519c762a2