Allow experienced users to claim mentees
This commit adds a special page Special:ClaimMentee,
which can be used by users who are listed on the
page with the list of mentors.

This special page allows mentors to change a mentor
of a mentee to themselves ("claim the mentee"),
provided the user is mentored by someone else
(has homepage enabled).

After a mentorship change is done, the mentee is
notified via Echo notification, if that extension
is installed.

Bug: T228210
Depends-On: I8505d2dad69aefed5f8ed1f599b3ca1b7e79acf5
Change-Id: I22f10a9632b50a52b372da0e274e32356e1e9071