Homepage: Adjust Mentor assignment in LocalUserCreated
$user->saveSettings() is not needed since the hook caller will do that. It's
possible that calling saveSettings twice in rapid succession (hook caller and
the deferred update in Mentor::saveMentor) caused the Special:Homepage to load
before mentorship assignment was saved in DB replica.

This patch also adds the user ID to logging when the mentor module can't be
rendered, so we can determine if it was an intermittent failure for the user.

Bug: T229235
Change-Id: Id7ff5f0979a979260a19558b70aa8bee92f7c047