Update posted on timestamp when overlay is opened
Listen to the mobileHomepageOverlayHtmlLoaded hook and parse the
timestamp data attribute, then update the display of the 'Posted $1 ago' string
when the help/mentorship overlays are opened.

There is a related but separate issue in that the most recent question is no
longer shown in the overlay immediately after posting a question, but that will
be a follow-up phab task.

Bug: T225941
Change-Id: I59e7a0c2140493d92ca7340a0e479435676d1be3