QuestionStore: Consider an answer archived if the URL has changed
An URL changing typically means a page move, which typically means
the question got archived. There are ways in which this can fail
(e.g. change in $wgArticlePath or fragment encoding rules) but they
are very rare edge cases and the code path triggered by the
condition check doesn't do anything irreversible so dealing with
those issues can be done when they arise.

Bug: T222208
Change-Id: Iced3f242d5b1429986584d88cc454a35e1c6be6c