Homepage: Unbreak question dialogs on mobile
Now that every module is in the page HTML in a hidden div, our code to
attach click handlers to the "ask a question" buttons is attaching them
to the hidden buttons, not the visible ones in the overlays.

Fix this by using the second $content parameter passed to the
OverlayHtmlLoaded hook to scope the button selectors.

Follow-up to 4aca04d7890013.

Bug: T229271
Change-Id: Ib440c91025fd841345e5df2a4d15a7a62b192e74
(cherry picked from commit 735a3ade227605be875d3fdc572232ef3bf5fd2d)