Homepage: Fix start module layout bugs
Remove rule in the mobile summary view that sets display: block; on a
div that's supposed to be display: flex;. This fix a layout bug in
Firefox (T230629).

Use flex-grow: 1; instead of height: 100%; to make the submodule bodies
grow to fill the available space when laid out horizontally (in the
desktop view). This fixes a layout bug in Safari 10 (T232549).

The display: block; rule was the fix for a different layout bug in iOS
Safari 10 (T255668), and removing it un-fixes this bug. However,
removing the height: 100%; rule also works to fix bug, so this change
does not regress it.

Follow-up to 6b461b1c3a10.

Bug: T230629
Bug: T232549
Bug: T225668
Change-Id: Idf015ed0263f6f1a16485b54053fe7d625ea880f
(cherry picked from commit 8dd04b5d7ce01de34722f8701ca400ab1b05aade)