Homepage: Remove question asking functionality from help module
Also remove a bunch of related code:
- QuestionPoster subclass
- Question tracking preference
- Guided tour
- Count of archived and unarchived questions in instrumentation data

Keep the tag for help module questions (because edits with this tag
already exist in the database), but mark it as inactive.

ApiQuestionStoreTest used the help module question poster for an
integration test, convert this to use the mentorship module question
poster instead.

Bug: T232548
Change-Id: Icdadc8f8f901c56224454f8eede250d2ebc66d12