Homepage: Add startediting placeholder module for suggested edits
If $wgGEHomepageSuggestedEditsEnabled is true, the userpage module is
replaced with a startediting module. This module displays different
messages depending on the user's WelcomeSurvey response.

When the button is clicked, a preference is set to indicate that the
user has activated suggested edits, and the module hides itself.

We'll still need to:
- Create the suggested edits module that gets added to the homepage when
  that happens
- Make this activation preference also rearrange the other modules
- Add interstitial overlays before the activation flag is set

Bug: T232419
Change-Id: I238c6ea2d520925d78f417224a0372e50fcc7400