Homepage: add banners for mobile discovery
Add mobile banners to help new users discover the homepage.
Since the situations in which to display the banner are the
exact same as no-JS discovery, that logic is being reused;
to keep the number of ResourceLoader modules down, the
existing module is also reused.

Also change SiteNoticeAfter hook handling to make sure later
hooks cannot add notices. Our notices will only be seen once
by the user so that's OK.

Bug: T224883
Change-Id: I56089762c44906b80153be6c97c232e60771854f