Add JS and dialog for startediting module
Add a dialog that is opened when the user clicks the button in the
startediting module. For now this dialog only has one panel (the
difficulty panel), but it has the infrastructure in place so we can
easily add the topic panel to it later.

When the user completes the dialog, we set a hidden preference to
remember that they've completed it, then reload the page to reveal a
new layout where the startup module only has three submodules.

In a follow-up patch, I intend to improve this so that a refresh isn't
needed and we make the changes in JS instead (which is easy now, but
will be harder once the refresh causes the new tasks module to appear
and the other modules to get reorganized).

Bug: T232419
Change-Id: I31ad394f00db79143940c7ff387f12c951e9ecd6