WelcomeSurvey: Fix saving results when group is overridden
The query parameter we used for overriding the survey group was ?group=,
but the form propagated it to the next request as '_group', so it wasn't
getting picked up for the purposes of defining which form fields exist.
This meant that if your user was assigned the group 'NONE', and you used
the ?group= override to get a survey, none of the survey data would be
saved. However, the _group value would be saved, so the second time around
it'd be honored and saving the survey would work.

This change fixes that by changing the override query parameter to
?_group=, so that the existing hidden form field feeds into it. This
does mean that for manual overrides, we'll now need to use ?_group=foo
instead of ?group=foo.

Bug: T233263
Change-Id: Ic9d39183b6a787155b399765b1256cc104a5ce1a