API module and service abstraction for suggested edits
Add a task suggestion service (with mock implementations
only) and an API for querying it. The architecture
allows for different task types (e.g. template-based,
category-based) and suggestion services (e.g. Cirrus,
ORES) in the future.

The ability to enumerate all task types (needed for API
documentation and eventually by the frontend config
widget) is left to a later patch.

Bug: T234426
Change-Id: Ie7b719ff351cd02180af0770fcc27b30e2a7b892
Depends-On: I9801647c39578a76be66f6e8c3c06c352660ca1e