Add RemoteSearchTaskSuggester
Uses CirrusSearch for suggestions, forwarding searches to a remote
API, and loading search configuration from a remote wiki page.
Intended for easy development environment setup.

Also renames GrowthExperimentsEditSuggester to
GrowthExperimentsTaskSuggester for consistency.

Filtering protected pages is left to the client side for now.
There is also no attempt to handle continuation or totals on the
remote API, as the initial version of the feature is limited to a
fixed number of tasks.

Example configuration:
$wgGENewcomerTasksRemoteApiUrl = 'https://cs.wikipedia.org/w/api.php';
$wgGENewcomerTasksRemoteConfigTitle = 'mw:Growth/Personalized_first_day'
    . '/Newcomer_tasks/Prototype/templates/cs.json';
(with the configuration coming from
https://www.mediawiki.org/w/index.php?oldid=3426874 )

Bug: T234426
Change-Id: I9b8c0a18bdc2adf5a24993f26d0055474c9edfbb
Depends-On: I959ba87f8e283da9549b6bfcefd338e60b00be8f
Depends-On: I83931c1a19c1cb747380e17a4aff3b1fadd5fcf3