Mobile homepage: Don't clone module DOM in overlays
This .clone() call was necessary when I changed this code to use the DOM
instead of config to transport the HTML, but we were still recreating
overlays from scratch every time they were opened. I made a mental note
to remove this .clone() call in the patch that introduced overlay
caching, but I forgot.

This fixes bugs with the suggested edits module in mobile overlay mode,
and also reduces the size of the DOM: modules that are opened in an
overlay will no longer be in the DOM twice, with the original being in
the hidden overlay-container div, and a copy in the actually-visible
mobile overlay, but instead they'll be moved from the hidden div to the
overlay.

Change-Id: I80a736cc997f57b7e16123368694bc9ce3d4a9f6