Homepage: Refactor StartEditing.js to not use global selectors
Just like in the JS for the other buttons that launch dialogs (e.g.
the mentorship module), we have to both look for the button in the
initial DOM, and register a hook to look in the mobile overlay DOMs as
they're created.

This fixes the "See suggested edits" button on mobile, which previously
wasn't opening the dialog.

Change-Id: Iff72d406e23174e485ffc3ebce797a72133f1c66