Homepage: Refactor SuggestedEditsModule to not use global selectors
Instead, have it look for the pager, previous and next elements in
this.$element. The caller can either pass in the existing
(server-rendered) DOM as config.$element, or pass in nothing and let the
module construct it.

Also make it look both at the initial DOM and in mobile overlays using a
hook. This has the effect that, on mobile, the task suggestions won't
be fetched and the UI won't be built out until the user navigates to the
suggested edits module. We could consider prepopulating it after the
page finishes loading.

Change-Id: Iaee706734adabc6baa111fe53f53ac1bca772fc2