Suggested Edits: Save task types to user preference
When the user interacts with the difficulty filters, store their filter
selection in a hidden preference as a JSON blob.

The name of the preference was left more open ended so that it can be
reused when topic matching is implemented.

The patch does not conditionally add the preference depending on whether the
feature flag for suggested edits is enabled, as that seems unnecessary.

Bug: T235042
Change-Id: Ie9f4e7e9584abdb55e8d71e34a276de66de2dc36