SuggestedEdits: Make filter button say "Easy, Medium difficulty"
Rather than just "Easy, Medium". Not specifically requested in T236841,
but implied in mockups there and elsewhere.

This also makes the qqq for the messages a bit easier, because now
there's one message for the button label, and three messages that get
comma-listed together as passed as a parameter to the first message.

Bug: T236841
Change-Id: I2e283fded0fbf4c8c1b44cdf209328ed8aca1512