Suggested Edits: Load sticky filter state into difficulty dialog
While the preferences are saved as the user interacts with the checkboxes and
the state is correctly reflected as the user opens/closes the dialog, on page
reload the checkboxes defaulted back to "easy" tasks. This patch loads the users
preferences for the filters and checks/unchecks boxes in the difficulty filters
based on those settings.

Bug: T235042
Change-Id: If66a1e0285d9a8120c888261b95f80da91e830c8