SuggestedEdits: Allow hiding behind a hidden preference
If $wgGEHomepageSuggestedEditsRequiresOptIn is set to true, a hidden
preference is created, and only users with that hidden preference
enabled get the suggested edits feature. If RequiresOptIn is set to
false, all users (that have the homepage enabled) get the suggested
edits feature.

To enable the hidden preference, run this in the browser console:
  new mw.Api().saveOption( 'growthexperiments-homepage-suggestededits', 1 )
      .done( function() { window.location.reload() } );

Bug: T236968
Change-Id: I6aef18f04680a5bc8e44dc467a7c07d6bbc5a219