StartEditingDialog: Implement intro panel and progress indicator
Adds a config var for info links for things that aren't task types
(creating a new article, and adding an image). Things that aren't task
types don't have info links in this UI, and for places where we do need
info links for them (e.g. the task explanation popup) we can add that
information to the on-wiki .json file that houses the rest of the task
type config.

Bug: T235723
Change-Id: Ic34523ecc40688a3c4702c4e85080d4527163c2e