Make RESTBase errors trigger the error screen + add URL override
Make sure that failed promises propagate up to the error handler
(except preload promises as there is no reason not to show the
successfully loaded current task just because the preload errored
out). Remove the getExtractAndUpdateQueue duplication on the
initial fetch to keep the code simpler. Also add a dev override
for using production RESTBase.

This is a rush job; some of the details of the error card do not
make sense in this situation when the current card could not be
loaded but the others could. Still better than not showing
anything, though.

Bug: T236522
Change-Id: I1ff691b29b08e86d69308f447e4a4cfbf2e9bdf3