Suggested Edits: Prevent concurrent requests when toggling filters
This patch aborts API requests made when there are multiple concurrent requests
to fetch tasks. This can happen when the user interacts with the
difficulty filters.

Additionally, the patch ignores the exception thrown in this case, because we
don't want to show an error card behind the overlay when the user is interacting
with the filters. The error card will still be shown for errors relating to
RESTBase failures.

A related change is that a search is not performed when the dialog first opens;
useful in case the user is looking at a task, opens the filters and then
cancels: they should still see the same task.

Finally, the patch removes an extraneous parameter to fetchTasks().

Change-Id: I17277df7666d8dc32febd15a5d1c853aaefa45ff