Sign pageview token with user ID
Change the GE homepage pageview token from 32 random characters
to 24 random characters + a 8-char signature tied to the user ID
(which is still pseudorandom so should be indistinguishable from
a statistical perspective, as long as the 160 vs 120 bit
probability space does not make any difference, which it should't
until 10^15 or so events).

This will be useful once the pageview token is used in such a way
that it might be accidentally shared to other users.

Change-Id: Id2477301849cd18da1aecc935561b72b44d7af47