Pass pageview token via task URL parameter, and use it in edit/help log
Add the pageview token (wgGEHomepagePageviewToken) as a 'clickid'
query parameter to task card link URLs, and on the task side use it
to override the session ID parameters in the EditAttemptStep and
HelpPanel schemas.

(Note that this slightly breaks the uniqueness assumption for the
session ID in those schemas - if the user opens multiple pages from
the same homepage request, e.g. by opening tasks in new tabs,
those will have the same session ID. This shouldn't conflict with
how those schemas are used.)

Also force oversampling of the EditAttemptStep schema, and add query
parameters to edits links on the page to propagate the session ID and
the oversampling flag.

To make tracking less obtrusive, hide the parameter from JS
after the page loads.

Bug: T238249
Change-Id: I08f673594351e1732932f05f1ba5c229e208e1a7
Depends-On: Ie1429280ec6895822b9594da2245e7cab5fed02a
Depends-On: Ib4586607b4acb69bdfbd6a8de0f320b90969209a
Depends-On: I4ede70f310a35c95b6eb9cc34cfcf2baa77e69ee
Depends-On: I300855fbf2aac016767582bcf40df0e369e6b862